Rename and clean up the Redriven Monitor Set class
